Create PFE.C7.E3.py
Making an easter egg in the program was fun. I ended up searching for all of the classical implementation of them and it was really entertaining.
To be honest this code I though that it had to do the same as the ex. 2 but for the examples runs in the book this is whit the count and the name of the file