Create PFE.C9.E3.py
The third exercise was sort of like the second one but instead of the count for day of the week is sorted by the email sender.
Using the .get() function for dictionaries has ben helpful to make the code smaller. I will probably use a if, else statement if I had not search for that function.