Create PFE.C13.3Coursera.py
This code is to retrieve a location, a plus code of a given location for the assignment in the coursera course.
Working with JSON is more straight forward and once I learned how to "pretty print" the data it was just a matter of figuring out how to extract the data needed.