Create PFE.C15.RoleCoursera.py
From the course the code was made to create a SQL database for students and their subjects alongside their association to the course.
It was surprising how much easier and intuitive the SQL programing can be, the only thing is to really know what to is there to be done.